Abstract away dependance of AngleGyroCorrection on PIDSource
